add error handles and optimization

+++ backend/routes/posts.js
@@ -1,38 +1,18 @@
 const express = require("express");
 const Post = require('../models/post');
-const multer = require('multer');
+
 const router = express.Router();
 
 const checkAuth = require("../middleware/check-auth");
+const extractFile = require("../middleware/file");
 
-const MIME_TYPE_MAP = {
-    'image/png': 'png',
-    'image/jpeg': 'jpeg',
-    'image/jpg': 'jpg'
-}
-
-const storage = multer.diskStorage({
-    destination: (req, file, cb) => {
-        const isValid = MIME_TYPE_MAP[file.mimetype];;
-        let error = new Error('invalid mime type');
-        if (isValid) {
-            error = null;
-        }
-        cb(error, "backend/images")
-    },
-    fileName: (req, file, cb) => {
-        const name = file.originalname.toLowerCase().split(" ").join("-");
-        const ext = MIME_TYPE_MAP[file.mimetype];
-        cb(null, name + '-' + Date.now() + '.' + ext);
-    }
-});
-
-router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+router.post("", checkAuth, extractFile, (req, res, next) => {
     const url = req.protocol + "://" + req.get("host");
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     })
     post.save().then(createdPost => {
         res.status(201).json({
@@ -45,8 +25,11 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
                 // imagePath: createdPost.imagePath
             }
         });
-    });
-
+    }).catch(error => {
+        res.status(500).json({
+            message: "creating a post failed!"
+        })
+    })
 })
 
 router.get("", (req, res, next) => {
@@ -66,7 +49,11 @@ router.get("", (req, res, next) => {
             posts: fechedPosts,
             maxPosts: count
         });
-    });
+    }).catch(error => {
+        res.status(500).json({
+            message: "Fetching post failed!"
+        })
+    })
 });
 
 router.get("/:id", (req, res, next) => {
@@ -76,10 +63,14 @@ router.get("/:id", (req, res, next) => {
         } else {
             res.status(404).json({message: "post not found"})
         }
+    }).catch(error => {
+        res.status(500).json({
+            message: "Fetching post failed!"
+        })
     })
 })
 
-router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+router.put("/:id", checkAuth, extractFile, (req, res, next) => {
     let imagePath = req.body.imagePath;
     if(req.file) {
         const url = req.protocol + "://" + req.get("host");
@@ -89,18 +80,33 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        console.log(result);
-        res.status(200).json({message: "update successful"});
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({message: "update successful"});
+        } else {
+            res.status(401).json({message: "Not authorized message"});
+        }  
+    }).catch(error => {
+        res.status(500).json({
+            message: "Couldn't update post!"
+        })
     })
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "post deleted" })
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: "post deleted" })
+        } else {
+            res.status(401).json({message: "Not authorized"});
+        }
+    }).catch(error => {
+        res.status(500).json({
+            message: "Fetching post failed!"
+        })
     })
 })
